Add reordering command to language and currency, change primary key type of currency and language

+++ PersonalFinanceApp.Application/Features/Currencies/Queries/GetCurrencyById/GetCurrencyByIdQuery.cs
@@ -3,12 +3,12 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using MediatR;
-using PersonalFinanceApp.Application.Features.Common;
+using PersonalFinanceApp.Application.Features.Currencies.Common;
 
 
 namespace PersonalFinanceApp.Application.Features.Currencies.Queries.GetCurrencyById;
 
 public class GetCurrencyByIdQuery : IRequest<CurrencyDto>
 {
-    public byte Id {get;set;}
+    public int Id {get;set;}
 }
