+++ PersonalFinanceApp.Application/Features/Expenditures/Queries/GetExpenditureById/ExpenditureDetailsDto.cs
@@ -12,7 +12,7 @@ public class ExpenditureDetailsDto
     public Guid AccountingDocumentId { get; set; }
     public byte[] RowVersion { get; set; } = Array.Empty<byte>();
     public DateOnly DocumentDate { get; set; }
-    public byte CurrencyId { get; set; }
+    public int CurrencyId { get; set; }
     public string? Description { get; set; }
