+++ PersonalFinanceApp.Application/Features/Incomes/Queries/GetIncomeById/IncomeDetailsDto.cs
@@ -10,7 +10,7 @@ public class IncomeDetailsDto
     public Guid AccountingDocumentId { get; set; }
     public byte[] RowVersion { get; set; } = Array.Empty<byte>();
     public DateOnly DocumentDate { get; set; }
-    public byte CurrencyId { get; set; }
+    public int CurrencyId { get; set; }
     public string? Description { get; set; }
 
     public List<AccountingEntryDto> IncomeLedgerAccountLines { get; set; } = new();
