+++ PersonalFinanceApp.Domain/Entities/DocumentTypeTranslation.cs
@@ -14,7 +14,7 @@ public class DocumentTypeTranslation
 
 
     // Foreign key to the related language
-    public byte LanguageId { get; private set; }
+    public int LanguageId { get; private set; }
     public Language Language { get; private set; } = null!;
